adding validation in models
